requirements.txt
astroid==2.4.2
autopep8==1.5.4
certifi==2020.12.5
click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
isort==5.5.2
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
pycodestyle==2.6.0
pylint==2.6.0
toml==0.10.1
Werkzeug==1.0.1
wrapt==1.12.1
python-pptx==0.6.19
python-docx==0.8.11
deep-translator==1.5.0



